Replace autotools with a standalone ./configure
This makes './compile; make' build rtptools without GNU
and without autoregenerating anything. The ./configure
script is a simple, hand-written, standalone shell script
accompanied with a set of trivial have-*.c check programs
and compat-*.c implementations.

Currently, we only check for err() and strtonum(),
and whether struct msghdr has msg_control; this is
a proof of concept - we need to go through the old
configure.ac and Makefile.am and trhrough all the sources
to make sure nothing is left behind.

The idea of replacing the GNU autohell with this
comes from http://mandoc.bsd.lv/cgi-bin/cvsweb/INSTALL