replace hsearch() with a simple list
Currently, rtpplay uses a hash table to remember the SSRCs we have seen
and their timestamps, to be used when playing according to the timestamps.
That's overkill: we don't need a full-fledged hash table implementation.
This diff replaces that with a simple list.

Then we don't need to check for a hsearch(3) implementation,
don't need to use the bundled buggy 1993 implementation (glibc 2.0.4),
and can get rid of one more dependency.